some modifications in (point check made in actions.h

+++ Actions/DeleteAction.cpp
@@ -12,13 +12,27 @@ void DeleteAction:: ReadActionParameters()
 	Output*pOut	=pManager->GetOutput();	
 	pOut->PrintMessage("Click on the figure you want to delete.");
 	Input*Input=pManager->GetInput();
-	Input->GetPointClicked(Delete.x, Delete.y); 
+	Input->GetPointClicked(Delete.x, Delete.y);
+	if(!Pointcheck(Delete))
+		Delete.y = Delete.x = 0;
 }
 void DeleteAction:: Execute() 
 {
+	Output*pOut	=pManager->GetOutput();
 	ReadActionParameters();
+
+	if(!Delete.y)
+	{
+		pOut->PrintMessage("Deleting has terminated, Choose another icon from the tool bar");
+		return;
+	}
+
 	CFigure* DelFigure = pManager->GetFigure(Delete.x, Delete.y);
 	pManager->DeleteFigure(DelFigure);
+	{
+		pOut->PrintMessage("A figure was deleted successfully, Choose another icon from the tool bar");
+	}
+	
 }
